+++ pkg/R/earthui-package.R
@@ -1,6 +1,7 @@
 #' @keywords internal
 "_PACKAGE"
 
+#' @import valengrCore
 #' @importFrom earth earth evimp
 #' @importFrom ggplot2 .data
 NULL